+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-deploy
 
+## 0.6.18
+
+### Patch Changes
+
+- deterministic deployment via create2 factory
+- Updated dependencies
+  - rocketh@0.6.16
+
 ## 0.6.17
 
 ### Patch Changes