+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-doc
 
+## 0.6.16
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.16
+
 ## 0.6.15
 
 ### Patch Changes