+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-verifier
 
+## 0.6.20
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.16
+
 ## 0.6.19
 
 ### Patch Changes